+++ author/js/author.js
@@ -133,6 +133,17 @@ function initAuthorCanvasser(vari, datafile, dataForm){
     else authorLibs.utils.requestJson(authorLibs.defaultJSON, initEdit);
   }
 
+  // link=document.createElement('link');
+  // link.href='https://fonts.googleapis.com/css?family=Indie+Flower';
+  // link.rel = 'stylesheet';
+  //
+  // document.getElementsByTagName('head')[0].appendChild(link);
+
+ var newcss = '@font-face {  font-family: testfont; src: url(./font/snoot-pixel10.woff);}';
+ var style = document.createElement('style');
+ style.appendChild(document.createTextNode(newcss));
+ document.getElementsByTagName('head')[0].appendChild(style);
+
   function initEdit(datafile){
     authorLibs.author = new authorcanvasser(datafile, 'file');
   }